pull out getserverside props
